# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseBodyError'
      status:
        type: integer
    type: object
  github_com_SanExpett_film-library-backend_internal_server_delivery.Response:
    properties:
      body:
        $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseBody'
      status:
        type: integer
    type: object
  github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseBody:
    properties:
      message:
        type: string
    type: object
  github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseBodyError:
    properties:
      error:
        type: string
    type: object
  github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseBodyID:
    properties:
      id:
        type: integer
    type: object
  github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseID:
    properties:
      body:
        $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseBodyID'
      status:
        type: integer
    type: object
  github_com_SanExpett_film-library-backend_pkg_models.Actor:
    properties:
      autor_id:
        type: integer
      birthday:
        type: string
      created_at:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_SanExpett_film-library-backend_pkg_models.ActorWithoutID:
    properties:
      birthday:
        type: string
      gender:
        type: string
      name:
        type: string
    type: object
...
  internal_actor_delivery.ActorListResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/github_com_SanExpett_film-library-backend_pkg_models.Actor'
        type: array
      status:
        type: integer
    type: object
  internal_actor_delivery.ActorResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_SanExpett_film-library-backend_pkg_models.Actor'
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a server of FILM-LIBRARY server.
  title: FILM-LIBRARY project API
  version: "1.0"
paths:
  /actor/add:
    post:
      consumes:
      - application/json
      description: |-
        add Actor by data
        Error.status can be:
        StatusErrBadRequest      = 400
        StatusErrInternalServer  = 500
      parameters:
      - description: Actor data for adding
        in: body
        name: Actor
        required: true
        schema:
          $ref: '#/definitions/github_com_SanExpett_film-library-backend_pkg_models.ActorWithoutID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseID'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: add Actor
      tags:
      - Actor
  /actor/delete:
    delete:
      consumes:
      - application/json
      description: |-
        delete Actor for author using user id from cookies\jwt.
        This totally removed Actor. Recovery will be impossible
      parameters:
      - description: Actor id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.Response'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: delete Actor
      tags:
      - Actor
  /actor/get:
    get:
      consumes:
      - application/json
      description: get Actor by id
      parameters:
      - description: Actor id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_actor_delivery.ActorResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get Actor
      tags:
      - Actor
  /actor/get_list_of_actors_in_film:
    get:
      consumes:
      - application/json
      description: get actors by film id
      parameters:
      - description: film id
        in: query
        name: film_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_actor_delivery.ActorListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get actors list starred in film
      tags:
      - Actor
  /actor/update:
    patch:
      consumes:
      - application/json
      description: update Actor by id
      parameters:
      - description: Actor id
        in: query
        name: id
        required: true
        type: integer
      - description: полностью опционален
        in: body
        name: preActor
        schema:
          $ref: '#/definitions/github_com_SanExpett_film-library-backend_pkg_models.ActorWithoutID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseID'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update Actor
      tags:
      - Actor
    put:
      consumes:
      - application/json
      description: update Actor by id
      parameters:
      - description: Actor id
        in: query
        name: id
        required: true
        type: integer
      - description: полностью опционален
        in: body
        name: preActor
        schema:
          $ref: '#/definitions/github_com_SanExpett_film-library-backend_pkg_models.ActorWithoutID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ResponseID'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_SanExpett_film-library-backend_internal_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update Actor
      tags:
      - Actor
...
schemes:
- http
swagger: "2.0"